Add possible release of ownership of hits
Add the possiblity to release the ownership of the hits. Now the trackpoint object does not need to own the hit objects. List of hits can be hold by a external structure and be shared between different tracks. An already existing track can release the hits, and load new list of hits for a new track fit procedure.