Correct outbound index of endId in getTOF
the endId index should not be incremented in the case of the function getTOF. If it is the default it is already -1, which because size-1 which is the last element of trackpoint vector. Adding +1 will provoke a outbound segmentation fault. Similarly if the endId is from the defined by user it has to be this trackpoint not the next one.